# wait_for_tasks: Execute own tasks first