--- a/src/lib/operators/aggregate_hash.cpp
+++ b/src/lib/operators/aggregate_hash.cpp
@@ -8,8 +8,6 @@
 #include <utility>
 #include <vector>
 
-#include <magic_enum.hpp>
-
 #include "aggregate/aggregate_traits.hpp"
 #include "constant_mappings.hpp"
 #include "expression/pqp_column_expression.hpp"