--- a/src/lib/scheduler/worker.hpp
+++ b/src/lib/scheduler/worker.hpp
@@ -2,6 +2,7 @@
 
 #include <atomic>
 #include <memory>
+#include <random>
 #include <thread>
 #include <vector>
 
@@ -66,6 +67,9 @@ class Worker : public std::enable_shared_from_this<Worker>, private Noncopyable 
   CpuID _cpu_id;
   std::thread _thread;
   std::atomic<uint64_t> _num_finished_tasks{0};
+
+  std::vector<int> _random{};
+  size_t _next_random{};
 };
 
 }  // namespace opossum