--- a/src/lib/scheduler/worker.hpp
+++ b/src/lib/scheduler/worker.hpp
@@ -2,7 +2,6 @@
 
 #include <atomic>
 #include <memory>
-#include <random>
 #include <thread>
 #include <vector>