(This PR replaces #2311. Something is amiss with Jenkins in a way that we have not seen before. Assuming that this is a freak accident, I simply reopened the PR. If this reoccurs, we have to look into it.)

Currently, `Worker::wait_for_tasks` checks if all tasks are completed. If they are not, the worker pulls tasks from the queue and executes them until all tasks that we wait on are done. This, however, leads to a problem if one of the pulled tasks calls `wait_for_tasks` itself. Over time, the worker builds a stack of `wait_for_tasks` calls and the first task has to wait until all tasks that the worker started because it was bored are completed.

<details>

<summary>Demo Code - click to expand</summary>

```cpp
int main() {
  const auto scheduler = std::make_shared<NodeQueueScheduler>();
  Hyrise::get().topology.use_non_numa_topology(3);
  Hyrise::get().set_scheduler(scheduler);

  // Start with some busy workers
  for (auto i = 0; i < 2; ++i) {
    std::make_shared<JobTask>([i] {
      sleep(i + 1);
      std::cout << "Adding one worker" << std::endl;
    })->schedule();
  }

  // Spawn 10 long-running tasks
  auto schedule_timer = Timer{};
  for (auto i = 0; i < 5; ++i) {
    std::make_shared<JobTask>([schedule_timer, i]() mutable {
      auto run_timer = Timer{};
      std::cout << std::string{"Started query "} + std::to_string(i) + "\n";

      auto a = std::make_shared<JobTask>([i] {
        std::cout << std::string{"Started left input "} + std::to_string(i) + "\n";
        sleep(i + 1);
        std::cout << std::string{"Finished left input "} + std::to_string(i) + "\n";
      });

      auto b = std::make_shared<JobTask>([i] {
        std::cout << std::string{"Started right input "} + std::to_string(i) + "\n";
        sleep(i + 1);
        std::cout << std::string{"Finished right input "} + std::to_string(i) + "\n";
      });

      auto c = std::make_shared<JobTask>([i] { std::cout << std::string{"Joined "} + std::to_string(i) + "\n"; });

      a->set_as_predecessor_of(c);
      b->set_as_predecessor_of(c);

      Hyrise::get().scheduler()->schedule_and_wait_for_tasks({a, b, c});

      std::cout << std::string{"Finished query "} + std::to_string(i) + " - " + schedule_timer.lap_formatted() +
                       " after being scheduled, execution took " + run_timer.lap_formatted() + "\n";
    })->schedule();
  }

  Hyrise::get().scheduler()->finish();

  return 0;
}
```
</details>

Results in the current master:
```
Finished query 4 - 13 s 1 ms after being scheduled, execution took 13 s 0 ms
Finished query 3 - 13 s 1 ms after being scheduled, execution took 13 s 0 ms
Finished query 2 - 13 s 1 ms after being scheduled, execution took 13 s 0 ms
Finished query 1 - 13 s 1 ms after being scheduled, execution took 13 s 0 ms
Finished query 0 - 13 s 1 ms after being scheduled, execution took 13 s 1 ms
```

With this PR:
```
Finished query 0 - 2 s 1 ms after being scheduled, execution took 2 s 0 ms
Finished query 1 - 5 s 1 ms after being scheduled, execution took 4 s 0 ms
Finished query 2 - 8 s 1 ms after being scheduled, execution took 6 s 0 ms
Finished query 3 - 10 s 1 ms after being scheduled, execution took 8 s 0 ms
Finished query 4 - 13 s 1 ms after being scheduled, execution took 8 s 0 ms
```

<details>
<summary>Full results</summary>

Master:
```
Started query 0
Started query 1
Started query 2
Started query 3
Started query 4
Started left input 0
Adding one worker
Started right input 0
Finished left input 0
Started left input 1
Adding one worker
Started right input 1
Finished right input 0
Joined 0
Started left input 2
Finished left input 1
Started right input 2
Finished right input 1
Joined 1
Started left input 3
Finished left input 2
Started right input 3
Finished right input 2
Joined 2
Started left input 4
Finished left input 3
Started right input 4
Finished right input 3
Joined 3
Finished left input 4
Finished right input 4
Joined 4
Finished query 4 - 13 s 1 ms after being scheduled, execution took 13 s 0 ms
Finished query 3 - 13 s 1 ms after being scheduled, execution took 13 s 0 ms
Finished query 2 - 13 s 1 ms after being scheduled, execution took 13 s 0 ms
Finished query 1 - 13 s 1 ms after being scheduled, execution took 13 s 0 ms
Finished query 0 - 13 s 1 ms after being scheduled, execution took 13 s 1 ms
```

New:
```
Started query 0
Started left input 0
Adding one worker
Started query 1
Started left input 1
Finished left input 0
Started right input 0
Adding one worker
Started query 2
Started left input 2
Finished right input 0
Joined 0
Finished query 0 - 2 s 1 ms after being scheduled, execution took 2 s 0 ms
Started query 3
Started left input 3
Finished left input 1
Started right input 1
Finished left input 2
Started right input 2
Finished right input 1
Joined 1
Finished query 1 - 5 s 1 ms after being scheduled, execution took 4 s 0 ms
Started query 4
Started left input 4
Finished left input 3
Started right input 3
Finished right input 2
Joined 2
Finished query 2 - 8 s 1 ms after being scheduled, execution took 6 s 0 ms
Started right input 4
Finished left input 4
Finished right input 3
Joined 3
Finished query 3 - 10 s 1 ms after being scheduled, execution took 8 s 0 ms
Finished right input 4
Joined 4
Finished query 4 - 13 s 1 ms after being scheduled, execution took 8 s 0 ms
```

</details>

This increases the performance for the analytical benchmarks quite a bit. For 200 clients (i.e., an overcommitted system), it also drastically reduces the latency.

TPC-C has a slight regression, but given that we have not looked into OLTP scheduling yet, I would keep that for another day. In a follow-up project, one could also look at how this affects the fairness. As we have no metrics for that yet, I would leave that for another day, too.

Also, this finally makes htop look like it is supposed to (node 0 is this branch, node 3 is master):

<img width="300" alt="Screenshot 2021-01-25 at 18 02 42" src="https://user-images.githubusercontent.com/575106/105767016-c9aa2b00-5f5a-11eb-8657-9df2cb3b5bd1.png">

===========

**System**
<details>
<summary>nemea - click to expand</summary>

| property | value |
| -- | -- |
| Hostname | nemea |
| CPU | Intel(R) Xeon(R) Platinum 8180 CPU @ 2.50GHz |
| Memory | 1.3TB |
| numactl | nodebind: 1  |
| numactl | membind: 1  |
</details>

**Commit Info and Build Time**
| commit | date | message | build time |
| -- | -- | -- | -- |
| 3b2f4c148 | 24.01.2021 12:16 | Update README.md (#2309) | real 447.40 user 2952.27 sys 142.88|
| c80fd32cb | 25.01.2021 21:59 | Change random engine | real 450.04 user 2939.22 sys 143.85|


**hyriseBenchmarkTPCH - single-threaded**
<details>
<summary>
Sum of avg. item runtimes: +1%
 ||
Geometric mean of throughput changes: +1%
</summary>

<details>
<summary>Configuration Overview - click to expand</summary>

```diff
 +Configuration Overview----+----------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------------------------------------------------------------------+
 | Parameter                | /home/Markus.Dreseler/hyrise4/build-release/benchmark_all_results/hyriseBenchmarkTPCH_3b2f4c1487617e8cffd00f0c1f888e4a5442165a_st.json | /home/Markus.Dreseler/hyrise4/build-release/benchmark_all_results/hyriseBenchmarkTPCH_c80fd32cbb795d17e5b0f492200a1e6ab60f19a6_st.json |
 +--------------------------+----------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------------------------------------------------------------------+
 |  GIT-HASH                | 3b2f4c1487617e8cffd00f0c1f888e4a5442165a-dirty                                                                                         | c80fd32cbb795d17e5b0f492200a1e6ab60f19a6-dirty                                                                                         |
 |  benchmark_mode          | Ordered                                                                                                                                | Ordered                                                                                                                                |
 |  build_type              | release                                                                                                                                | release                                                                                                                                |
 |  chunk_size              | 65535                                                                                                                                  | 65535                                                                                                                                  |
 |  clients                 | 1                                                                                                                                      | 1                                                                                                                                      |
 |  compiler                | gcc 10.2                                                                                                                               | gcc 10.2                                                                                                                               |
 |  cores                   | 0                                                                                                                                      | 0                                                                                                                                      |
 |  date                    | 2021-01-26 02:14:15                                                                                                                    | 2021-01-26 10:16:33                                                                                                                    |
 |  encoding                | {'default': {'encoding': 'Dictionary'}}                                                                                                | {'default': {'encoding': 'Dictionary'}}                                                                                                |
 |  indexes                 | False                                                                                                                                  | False                                                                                                                                  |
 |  max_duration            | 60000000000                                                                                                                            | 60000000000                                                                                                                            |
 |  max_runs                | -1                                                                                                                                     | -1                                                                                                                                     |
 |  scale_factor            | 1.0                                                                                                                                    | 1.0                                                                                                                                    |
 |  time_unit               | ns                                                                                                                                     | ns                                                                                                                                     |
 |  use_prepared_statements | False                                                                                                                                  | False                                                                                                                                  |
 |  using_scheduler         | False                                                                                                                                  | False                                                                                                                                  |
 |  verify                  | False                                                                                                                                  | False                                                                                                                                  |
 |  warmup_duration         | 0                                                                                                                                      | 0                                                                                                                                      |
 +--------------------------+----------------------------------------------------------------------------------------------------------------------------------------+----------------------------------------------------------------------------------------------------------------------------------------+
```
</details>

```diff
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Item     || Latency (ms/iter)  | Change || Throughput (iter/s) | Change | p-value |
 |          ||      old |     new |        ||      old |      new |        |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | TPC-H 01 ||    518.5 |   509.1 |   -2%  ||     1.93 |     1.96 |   +2%  |  0.0037 |
+| TPC-H 02 ||      4.2 |     3.9 |   -7%  ||   235.65 |   253.42 |   +8%  |  0.0000 |
 | TPC-H 03 ||     71.4 |    70.7 |   -1%  ||    14.01 |    14.14 |   +1%  |  0.0000 |
 | TPC-H 04 ||     63.3 |    61.9 |   -2%  ||    15.79 |    16.17 |   +2%  |  0.0000 |
 | TPC-H 05 ||    194.1 |   198.7 |   +2%  ||     5.15 |     5.03 |   -2%  |  0.0000 |
 | TPC-H 06 ||      3.4 |     3.4 |   -2%  ||   292.14 |   296.93 |   +2%  |  0.0000 |
 | TPC-H 07 ||     50.7 |    50.7 |   +0%  ||    19.73 |    19.72 |   -0%  |  0.9736 |
 | TPC-H 08 ||     54.7 |    55.2 |   +1%  ||    18.27 |    18.11 |   -1%  |  0.1561 |
 | TPC-H 09 ||    442.3 |   447.5 |   +1%  ||     2.26 |     2.23 |   -1%  |  0.0010 |
 | TPC-H 10 ||    102.4 |   102.3 |   -0%  ||     9.76 |     9.78 |   +0%  |  0.4459 |
+| TPC-H 11 ||      8.1 |     7.6 |   -7%  ||   123.41 |   132.13 |   +7%  |  0.0000 |
 | TPC-H 12 ||     41.0 |    40.6 |   -1%  ||    24.36 |    24.61 |   +1%  |  0.0000 |
 | TPC-H 13 ||    329.5 |   335.1 |   +2%  ||     3.03 |     2.98 |   -2%  |  0.0000 |
 | TPC-H 14 ||     17.8 |    18.1 |   +1%  ||    56.16 |    55.35 |   -1%  |  0.0000 |
 | TPC-H 15 ||      6.8 |     6.9 |   +1%  ||   147.06 |   145.75 |   -1%  |  0.0000 |
 | TPC-H 16 ||     63.2 |    63.9 |   +1%  ||    15.82 |    15.64 |   -1%  |  0.0000 |
+| TPC-H 17 ||     15.4 |    14.6 |   -5%  ||    65.13 |    68.70 |   +5%  |  0.0000 |
-| TPC-H 18 ||    503.4 |   528.4 |   +5%  ||     1.99 |     1.89 |   -5%  |  0.0000 |
 | TPC-H 19 ||     27.8 |    27.1 |   -2%  ||    36.03 |    36.94 |   +3%  |  0.0000 |
 | TPC-H 20 ||     14.3 |    13.9 |   -3%  ||    70.05 |    71.91 |   +3%  |  0.0000 |
 | TPC-H 21 ||    337.3 |   336.8 |   -0%  ||     2.96 |     2.97 |   +0%  |  0.7048 |
 | TPC-H 22 ||     31.3 |    31.0 |   -1%  ||    31.94 |    32.24 |   +1%  |  0.0000 |
 +----------++----------+---------+--------++----------+----------+--------+---------+
 | Sum      ||   2900.9 |  2927.3 |   +1%  ||          |          |        |         |
 | Geomean  ||          |         |        ||          |          |   +1%  |         |
 +----------++----------+---------+--------++----------+----------+--------+---------+
```
</details>


**hyriseBenchmarkTPCH - single-threaded, SF 0.01**
<details>
<summary>
Sum of avg. item runtimes: +1%
 ||
Geometric mean of throughput changes: -1%
</summary>

<details>
<summary>Configuration Overview - click to expand</summary>

```diff
 +Configuration Overview----+--------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+
 | Parameter                | /home/Markus.Dreseler/hyrise4/build-release/benchmark_all_results/hyriseBenchmarkTPCH_3b2f4c1487617e8cffd00f0c1f888e4a5442165a_st_s01.json | /home/Markus.Dreseler/hyrise4/build-release/benchmark_all_results/hyriseBenchmarkTPCH_c80fd32cbb795d17e5b0f492200a1e6ab60f19a6_st_s01.json |
 +--------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+--------------------------------------------------------------------------------------------------------------------------------------------+
 |  GIT-HASH                | 3b2f4c1487617e8cffd00f0c1f888e4a5442165a-dirty                                                                                             | c80fd32cbb795d17e5b0f492200a1e6ab60f19a6-dirty                                                                                             |
 |  benchmark_mode          | Ordered                                                                                                                                    | Ordered                                                                                                                                    |
 |  build_type              | release                                                                                                                                    | release                                                                                                                                    |
 |  chunk_size              | 65535                                                                                                                                      | 65535                                                                                                                                      |
 |  clients                 | 1                                                                                                                                          | 1                                                                                                                                          |
 |  compiler                | gcc 10.2                                                                                                                                   | gcc 10.2                                                                                                                                   |
 |  cores                   | 0                                                                                                                                          | 0                                                                                                                                          |
…