@Bouncner 2021-02-15
> The results in the description are current. 

Hmm? Looks like it does not include the random values in a vector during construction. So that part that brought TPC-C back.