@Bouncner 2021-02-15
> The right side is [72331c3](https://github.com/hyrise/hyrise/commit/72331c3a86447303c4f45d935ec2f7a7162d7921), you can find the full worker.cpp [here](https://github.com/hyrise/hyrise/blob/72331c3a86447303c4f45d935ec2f7a7162d7921/src/lib/scheduler/worker.cpp). It appears as if the changes to the random generation were not at the core of the issue. :/

Strange, it helped a lot when I tested it with a much worse approach. :-/